fix for functions
