+++ firebase/functions/src/index.ts
@@ -1,11 +1,20 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-
 import admin = require("firebase-admin")
-import * as serviceAccount from './serviceAccountKey.json';
-// import tts = require('./tts');
+import express = require("express")
+import { https } from "firebase-functions/v2"
+import { getTTS } from "./deepgram"
 
-admin.initializeApp({
-  credential: admin.credential.cert(serviceAccount as admin.ServiceAccount)}
-);
+admin.initializeApp()
+const app = express()
 
-exports.api = require('./tts')
+app.post("/audio/tts", async (req, res) => {
+  try {
+    const { text } = JSON.parse(req.body)
+    res.status(200).send(await getTTS(text))
+  } catch (error) {
+    res.status(500).send("Cudnt get tts rip")
+  }
+})
+
+// TODO: add endpoint for STT
+
+exports.api = https.onRequest(app)
